requirements.txt
amqp==1.4.7
anyjson==0.3.3
appnope==0.1.0
billiard==3.3.0.20
celery==3.1.18
decorator==4.0.4
Django==1.8.4
django-datetime-widget==0.9.3
django-social-auth==0.7.28

google-api-python-client==1.4.2
httplib2==0.9.2
ipython==4.0.0
ipython-genutils==0.1.0
kombu==3.0.28
oauth2==1.9.0.post1
oauth2client==1.5.1
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.5
ptyprocess==0.5
pyasn1==0.1.9
pyasn1-modules==0.0.8
pyreadline==2.1
python-dateutil==2.4.2
python-openid==2.2.5
pytz==2015.6
requests==2.8.0
rsa==3.2
simplegeneric==0.8.1
simplejson==3.8.0
six==1.10.0
traitlets==4.0.0
Unipath==1.1
uritemplate==0.6
wheel==0.24.0
django-filer==0.9.12
easy-thumbnails==2.2